Added some examples.
In the state machine, which is called on each received 9P message, we check
for a :mutation-stream key associated to the state map, passed in via the
initial server! call. If it exists, we fetch from it, and pass it in to a
function on each iteration. Since the value is a manifold deferred, we check
if it is realized. If it is, then we deserialise the values and execute a
function callback on the state, mutating it.

The messages on the stream have been defined as a map,

`:fn`, containing a function to execute, which takes as keyword arguments
`:state` and `:data`, where `:state` should be the current state, and `:data`
the modifying input for the funcall. The function can then execute and return
a modified state, which is then passed as normal back through the regular
flow.

We use this to add a file to the virtual filesystem in the example.

Once the callback is executed, the deferred is dissociated from the state
map, and a new deferred is fetched, and awaiting another message to arrive
before executing.