Starting to separate the VFS / state coupling.
We can use the type/dev fields in a stat to also choose how to handle
the different types of files. Now that the proof-of-concept in-memory
hack of a "filesystem" is functional, I'm going to simplify the design
and try and make it more pluggable. To this end, I shall likely start
with a sqlite-backed store. By supporting more than one mode of access,
it'll probably highlight obvious deficiencies and pave the way for a
somewhat more sensible design.