Some extensions to the tcp routing mechanism.
This is some initial work to see if there is some nice developer-friendly
interface that can be created to allow greater ergonomics for custom handler
implementations. Right now almost the whole server namespace, along with the
state handler, etc., need to be essentially duplicated in real-world
implementations. The frame handlers, assembly/disassembly, etc. are the only
real static out-of-the-box drop-in things that I am happy with. Herein lies the
beginning of making this an easier library for direct integration into other
upstream projects. (Admittedly, primarily my own for starters.)